+++ src/constraints/cartesian_product.rs
@@ -22,12 +22,12 @@ use super::Constraint;
 /// for all $i=0,\ldots, n-1$.
 ///
 #[derive(Default)]
-pub struct CartesianProduct<'a> {
+pub struct CartesianProduct<'a, T = f64> {
     idx: Vec<usize>,
-    constraints: Vec<Box<dyn Constraint + 'a>>,
+    constraints: Vec<Box<dyn Constraint<T> + 'a>>,
 }
 
-impl<'a> CartesianProduct<'a> {
+impl<'a, T> CartesianProduct<'a, T> {
     /// Construct a new Cartesian product of constraints.
     ///
     /// # Note
@@ -35,6 +35,21 @@ impl<'a> CartesianProduct<'a> {
     /// The use of `new_with_capacity` should be preferred over this method,
     /// when possible (provided you have an estimate of the number of sets
     /// your Cartesian product will consist of).
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// use optimization_engine::constraints::{Ball2, CartesianProduct, Constraint, Rectangle};
+    ///
+    /// let xmin = [-1.0, -1.0];
+    /// let xmax = [1.0, 1.0];
+    /// let cartesian = CartesianProduct::new()
+    ///     .add_constraint(2, Rectangle::new(Some(&xmin), Some(&xmax)))
+    ///     .add_constraint(4, Ball2::new(None, 1.0));
+    ///
+    /// let mut x = [3.0, -2.0, 2.0, 0.0];
+    /// cartesian.project(&mut x);
+    /// ```
     ///
     pub fn new() -> Self {
         CartesianProduct {
@@ -123,7 +138,8 @@ impl<'a> CartesianProduct<'a> {
     /// ```
     /// The method will panic if any of the associated projections panics.
     ///
-    pub fn add_constraint(mut self, ni: usize, constraint: impl Constraint + 'a) -> Self {
+    #[must_use]
+    pub fn add_constraint(mut self, ni: usize, constraint: impl Constraint<T> + 'a) -> Self {
         assert!(
             self.dimension() < ni,
             "provided index is smaller than or equal to previous index, or zero"
@@ -134,7 +150,7 @@ impl<'a> CartesianProduct<'a> {
     }
 }
 
-impl<'a> Constraint for CartesianProduct<'a> {
+impl<'a, T> Constraint<T> for CartesianProduct<'a, T> {
     /// Project onto the Cartesian product of constraints.
     ///
     /// The given vector `x` is updated with the projection on the set
@@ -143,7 +159,7 @@ impl<'a> Constraint for CartesianProduct<'a> {
     ///
     /// The method will panic if the dimension of `x` is not equal to the
     /// dimension of the Cartesian product (see `dimension()`)
-    fn project(&self, x: &mut [f64]) {
+    fn project(&self, x: &mut [T]) {
         assert!(x.len() == self.dimension(), "x has wrong size");
         let mut j = 0;
         self.idx