+++ src/core/fbs/fbs_engine.rs
@@ -5,25 +5,28 @@ use crate::{
     core::{fbs::FBSCache, AlgorithmEngine, Problem},
     matrix_operations, FunctionCallResult, SolverError,
 };
+use num::Float;
 
 /// The FBE Engine defines the steps of the FBE algorithm and the termination criterion
 ///
-pub struct FBSEngine<'a, GradientType, ConstraintType, CostType>
+pub struct FBSEngine<'a, GradientType, ConstraintType, CostType, T = f64>
 where
-    GradientType: Fn(&[f64], &mut [f64]) -> FunctionCallResult,
-    CostType: Fn(&[f64], &mut f64) -> FunctionCallResult,
-    ConstraintType: constraints::Constraint,
+    T: Float,
+    GradientType: Fn(&[T], &mut [T]) -> FunctionCallResult,
+    CostType: Fn(&[T], &mut T) -> FunctionCallResult,
+    ConstraintType: constraints::Constraint<T>,
 {
-    pub(crate) problem: Problem<'a, GradientType, ConstraintType, CostType>,
-    pub(crate) cache: &'a mut FBSCache,
+    pub(crate) problem: Problem<'a, GradientType, ConstraintType, CostType, T>,
+    pub(crate) cache: &'a mut FBSCache<T>,
 }
 
-impl<'a, GradientType, ConstraintType, CostType>
-    FBSEngine<'a, GradientType, ConstraintType, CostType>
+impl<'a, GradientType, ConstraintType, CostType, T>
+    FBSEngine<'a, GradientType, ConstraintType, CostType, T>
 where
-    GradientType: Fn(&[f64], &mut [f64]) -> FunctionCallResult,
-    CostType: Fn(&[f64], &mut f64) -> FunctionCallResult,
-    ConstraintType: constraints::Constraint,
+    T: Float,
+    GradientType: Fn(&[T], &mut [T]) -> FunctionCallResult,
+    CostType: Fn(&[T], &mut T) -> FunctionCallResult,
+    ConstraintType: constraints::Constraint<T>,
 {
     /// Constructor for instances of `FBSEngine`
     ///
@@ -36,13 +39,13 @@ where
     ///
     /// An new instance of `FBSEngine`
     pub fn new(
-        problem: Problem<'a, GradientType, ConstraintType, CostType>,
-        cache: &'a mut FBSCache,
-    ) -> FBSEngine<'a, GradientType, ConstraintType, CostType> {
+        problem: Problem<'a, GradientType, ConstraintType, CostType, T>,
+        cache: &'a mut FBSCache<T>,
+    ) -> FBSEngine<'a, GradientType, ConstraintType, CostType, T> {
         FBSEngine { problem, cache }
     }
 
-    fn gradient_step(&mut self, u_current: &mut [f64]) {
+    fn gradient_step(&mut self, u_current: &mut [T]) {
         assert_eq!(
             Ok(()),
             (self.problem.gradf)(u_current, &mut self.cache.work_gradient_u),
@@ -53,20 +56,21 @@ where
         u_current
             .iter_mut()
             .zip(self.cache.work_gradient_u.iter())
-            .for_each(|(u, w)| *u -= self.cache.gamma * *w);
+            .for_each(|(u, w)| *u = *u - self.cache.gamma * *w);
     }
 
-    fn projection_step(&mut self, u_current: &mut [f64]) {
+    fn projection_step(&mut self, u_current: &mut [T]) {
         self.problem.constraints.project(u_current);
     }
 }
 
-impl<'a, GradientType, ConstraintType, CostType> AlgorithmEngine
-    for FBSEngine<'a, GradientType, ConstraintType, CostType>
+impl<'a, GradientType, ConstraintType, CostType, T> AlgorithmEngine<T>
+    for FBSEngine<'a, GradientType, ConstraintType, CostType, T>
 where
-    GradientType: Fn(&[f64], &mut [f64]) -> FunctionCallResult + 'a,
-    CostType: Fn(&[f64], &mut f64) -> FunctionCallResult + 'a,
-    ConstraintType: constraints::Constraint + 'a,
+    T: Float,
+    GradientType: Fn(&[T], &mut [T]) -> FunctionCallResult + 'a,
+    CostType: Fn(&[T], &mut T) -> FunctionCallResult + 'a,
+    ConstraintType: constraints::Constraint<T> + 'a,
 {
     /// Take a forward-backward step and check whether the algorithm should terminate
     ///
@@ -83,7 +87,7 @@ where
     ///
     /// The method may panick if the computation of the gradient of the cost function
     /// or the cost function panics.
-    fn step(&mut self, u_current: &mut [f64]) -> Result<bool, SolverError> {
+    fn step(&mut self, u_current: &mut [T]) -> Result<bool, SolverError> {
         self.cache.work_u_previous.copy_from_slice(u_current); // cache the previous step
         self.gradient_step(u_current); // compute the gradient
         self.projection_step(u_current); // project
@@ -93,7 +97,7 @@ where
         Ok(self.cache.norm_fpr > self.cache.tolerance)
     }
 
-    fn init(&mut self, _u_current: &mut [f64]) -> FunctionCallResult {
+    fn init(&mut self, _u_current: &mut [T]) -> FunctionCallResult {
         Ok(())
     }
 }