+++ website/sidebars.js
@@ -14,6 +14,7 @@ module.exports = {
         'python-c',
         'python-bindings',
         'python-tcp-ip',
+        'python-ros2',
         'python-ros',
         'python-examples',
       ],
@@ -26,7 +27,7 @@ module.exports = {
     {
       type: 'category',
       label: 'Rust',
-      items: ['openrust-basic', 'openrust-alm', 'openrust-features'],
+      items: ['openrust-basic', 'openrust-alm', 'openrust-features', 'openrust-arithmetic'],
     },
     {
       type: 'category',